+++ src/main/java/com/supermartijn642/formations/structure/PiecedStructure.java
@@ -7,6 +7,7 @@ import com.supermartijn642.formations.FormationsStructures;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.LevelHeightAccessor;
 import net.minecraft.world.level.block.Rotation;
 import net.minecraft.world.level.chunk.ChunkGenerator;
@@ -90,7 +91,7 @@ public class PiecedStructure extends Structure {
         StructureTemplateManager templateManager = context.structureTemplateManager();
         LevelHeightAccessor levelHeightAccessor = context.heightAccessor();
         WorldgenRandom random = context.random();
-        Registry<StructureTemplatePool> registry = context.registryAccess().registryOrThrow(Registry.TEMPLATE_POOL_REGISTRY);
+        Registry<StructureTemplatePool> registry = context.registryAccess().registryOrThrow(Registries.TEMPLATE_POOL);
 
         PoolElementStructurePiece structurePiece = new PoolElementStructurePiece(templateManager, startElement, pos, startElement.getGroundLevelDelta(), rotation, startElement.getBoundingBox(templateManager, pos, rotation));
         BoundingBox boundingBox = structurePiece.getBoundingBox();