+++ changelog.md
@@ -1,2 +1,2 @@
 ### Formations 1.0.0
-- Initial release of Formations for Fabric on 1.19.2
+- Initial release of Formations for Fabric on 1.19.4